backend/requirements.txt
# Core FastAPI and server
# MCP SDK requires starlette>=0.27 and uvicorn>=0.31, so we use FastAPI 0.115+
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-multipart>=0.0.9

# Database (SQLite — zero-install, replaces MongoDB)
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
pyjwt[crypto]>=2.10.1
bcrypt>=4.0.1

# Settings and validation
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.5.2

# HTTP client
httpx>=0.27.0

# Encryption
cryptography>=41.0.7

# Search
ddgs>=9.0.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
numpy>=1.26.0,<2.0.0

# Vector store
chromadb>=1.4.1

# Entity extraction (GLiNER — used for query-time entity linking in graph search)
gliner>=0.2.24

# Neo4j graph database
neo4j>=5.17.0

# MCP Server (Windsurf integration)
mcp>=1.26.0
pywin32>=310; sys_platform == "win32"

# Document parsing (RAG)
pypdf>=4.0.0
python-docx>=1.1.0

# Date parsing (flexible datetime from LLM output)
python-dateutil>=2.8.2

# Web crawling / page fetching
diskcache>=5.6.0
markdownify>=0.13.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.0.0
